feat(#82): Scheduler, run history & ingest terminal — migrate to routed views
- SchedulerPage: interval presets + custom input, pause/resume, fetch now, live
  countdown; all feedback via sonner toasts (no more inline banners)
- FeedsRunsPage: paginated run history table with expandable per-source breakdown,
  using shadcn Table + Button components
- FeedsLogsPage: SSE terminal (EventSource /api/ingest/stream), theme-aware dark
  monospace pre, live/connecting status badge
- Remove IngestTerminal, RunHistoryTable from App.tsx; strip SchedulerTab to
  stats-only (controls now in SchedulerPage; stats remain for slice #83)
- Add formatDateTime, formatDuration, formatInteger, relativeCountdown to lib/format.ts
- Delete ~620 lines of scheduler/run-history/terminal CSS from styles.css
